fix: remove all circular dependencies
