+++ src/trace/spancontext-utils.ts
@@ -13,20 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { INVALID_SPANID, INVALID_TRACEID } from './invalid-span-constants';
 import { NonRecordingSpan } from './NonRecordingSpan';
 import { Span } from './span';
 import { SpanContext } from './span_context';
-import { TraceFlags } from './trace_flags';
 
 const VALID_TRACEID_REGEX = /^([0-9a-f]{32})$/i;
 const VALID_SPANID_REGEX = /^[0-9a-f]{16}$/i;
-export const INVALID_SPANID = '0000000000000000';
-export const INVALID_TRACEID = '00000000000000000000000000000000';
-export const INVALID_SPAN_CONTEXT: SpanContext = {
-  traceId: INVALID_TRACEID,
-  spanId: INVALID_SPANID,
-  traceFlags: TraceFlags.NONE,
-};
 
 export function isValidTraceId(traceId: string): boolean {
   return VALID_TRACEID_REGEX.test(traceId) && traceId !== INVALID_TRACEID;
